Refelcted comments and made changes
