+++ examples/speaker_tasks/diarization/conf/inference/diar_infer_telephonic.yaml
@@ -52,8 +52,8 @@ diarizer:
       max_rp_threshold: 0.25 # Determines the range of p-value search: 0 < p <= max_rp_threshold. 
       sparse_search_volume: 30 # The higher the number, the more values will be examined with more time. 
       maj_vote_spk_count: False  # If True, take a majority vote on multiple p-values to estimate the number of speakers.
-      sub_cluster_n: 50 # Number of sub-clusters for each sub-window for long-form audio clustering.
-      unit_window_len: 10000 # Window length for the unit of chunking. Control according to the GPU-memory capacity.
+      chunk_cluster_count: 50 # Number of forced clusters (over-clustering) per unit chunk in long-form audio clustering.
+      embeddings_per_chunk: 10000 # Number of embeddings in each chunk for long-form audio clustering. Adjust based on GPU memory capacity. 
   
   msdd_model:
     model_path: diar_msdd_telephonic # .nemo local model path or pretrained model name for multiscale diarization decoder (MSDD)
