+++ nemo/collections/asr/parts/utils/offline_clustering.py
@@ -402,7 +402,6 @@ def get_argmin_mat(timestamps_in_scales: List[torch.Tensor]) -> List[torch.Tenso
     for scale_idx in scale_list:
         time_stamps_float = timestamps_in_scales[scale_idx]
         segment_anchor_list.append(torch.mean(time_stamps_float, dim=1))
-
     base_scale_idx = max(scale_list)
     base_scale_anchor = segment_anchor_list[base_scale_idx]
     session_scale_mapping_list = []
@@ -781,7 +780,7 @@ class SpectralClustering:
 
         Returns:
             labels (Tensor):
-                clustering label output
+                Clustering label output
         """
         if X.shape[0] != X.shape[1]:
             raise ValueError("The affinity matrix is not a square matrix.")
@@ -1109,7 +1108,6 @@ class NMESC:
             raise ValueError("p_value_list should not be empty.")
         return p_value_list
 
-
 class SpeakerClustering(torch.nn.Module):
     def __init__(
         self,
@@ -1119,7 +1117,6 @@ class SpeakerClustering(torch.nn.Module):
         maj_vote_spk_count: bool = False,
         parallelism: bool = False,
         cuda: bool = False,
-        device=None,
     ):
         """
         Clustering method for speaker diarization based on cosine similarity.
@@ -1130,6 +1127,8 @@ class SpeakerClustering(torch.nn.Module):
                 The minimum number of samples required for NME clustering. This avoids
                 zero p_neighbour_lists. If the input has fewer segments than min_samples,
                 it is directed to the enhanced speaker counting mode.
+            nme_mat_size (int):
+                The targeted matrix size for NME analysis.
             sparse_search (bool):
                 Toggle sparse search mode. If True, limit the size of p_value_list to sparse_search_volume.
             maj_vote_spk_count (bool):
@@ -1150,10 +1149,87 @@ class SpeakerClustering(torch.nn.Module):
         self.maj_vote_spk_count: bool = maj_vote_spk_count
         self.embeddings_in_scales: List[torch.Tensor] = [torch.Tensor(0)]
         self.timestamps_in_scales: List[torch.Tensor] = [torch.Tensor(0)]
-        if device is None or type(device) is not torch.device:
-            self.device = torch.device("cuda") if self.cuda else torch.device("cpu")
+        self.device = torch.device("cuda") if self.cuda else torch.device("cpu")
+            
+    def forward_unit_infer(
+        self,
+        mat: torch.Tensor,
+        oracle_num_speakers: int = -1,
+        max_num_speakers: int = 8,
+        max_rp_threshold: float = 0.15,
+        sparse_search_volume: int = 30,
+        est_num_of_spk_enhanced: torch.Tensor = torch.tensor(-1), 
+        fixed_thres: float = -1.0,
+        kmeans_random_trials: int = 1,
+    ) -> torch.LongTensor:
+        """
+        This function takes a cosine similarity matrix `mat` and returns the speaker labels for the segments 
+        in the given input embeddings. 
+       
+        Args: 
+            mat (Tensor):
+                Cosine similarity matrix (affinity matrix) calculated from the provided speaker embeddings.
+            oracle_num_speakers (int):
+                The number of speakers in a session, as specified by the reference transcript.
+                Can be used as `chunk_cluster_count` in long-form clustering mode.
+            max_num_speakers (int):
+                The upper bound for the number of speakers in each session.
+            max_rp_threshold (float):
+                Limits the range of parameter search.
+                The clustering performance can vary based on this range.
+                The default value is 0.15.
+            sparse_search_volume (int):
+                The number of p_values considered during NME analysis.
+                The default is 30. Lower values speed up the NME-analysis but might lead to poorer parameter estimations. Values below 20 are not recommended.
+            est_num_of_spk_enhanced (int):
+                The number of speakers estimated from enhanced speaker counting.
+                If the value is -1, the enhanced speaker counting is skipped.
+            fixed_thres (float):
+                If a `fixed_thres` value is provided, the NME-analysis process will be skipped.
+                This value should be optimized on a development set for best results.
+                By default, it is set to -1.0, and the function performs NME-analysis to estimate the threshold.
+            kmeans_random_trials (int):
+                The number of random trials for initializing k-means clustering. More trials can result in more stable clustering. The default is 1. 
+                
+        Returns:
+            Y (LongTensor):
+                Speaker labels (clustering output) in integer format for the segments in the given input embeddings.
+        """
+        nmesc = NMESC(
+            mat,
+            max_num_speakers=max_num_speakers,
+            max_rp_threshold=max_rp_threshold,
+            sparse_search=self.sparse_search,
+            sparse_search_volume=sparse_search_volume,
+            fixed_thres=fixed_thres,
+            nme_mat_size=self.nme_mat_size,
+            maj_vote_spk_count=self.maj_vote_spk_count,
+            parallelism=self.parallelism,
+            cuda=self.cuda,
+            device=self.device,
+        )
+        # If there are less than `min_samples_for_nmesc` segments, est_num_of_spk is 1.
+        if mat.shape[0] > self.min_samples_for_nmesc:
+            est_num_of_spk, p_hat_value = nmesc.forward()
+            affinity_mat = getAffinityGraphMat(mat, p_hat_value)
         else:
-            self.device = device
+            nmesc.fixed_thres = max_rp_threshold
+            est_num_of_spk, p_hat_value = nmesc.forward()
+            affinity_mat = mat
+
+        # n_clusters is number of speakers estimated from spectral clustering.
+        if oracle_num_speakers > 0:
+            n_clusters = int(oracle_num_speakers)
+        elif est_num_of_spk_enhanced > 0:
+            n_clusters = int(est_num_of_spk_enhanced.item())
+        else:
+            n_clusters = int(est_num_of_spk.item())
+
+        spectral_model = SpectralClustering(
+            n_clusters=n_clusters, n_random_trials=kmeans_random_trials, cuda=self.cuda, device=self.device
+        )
+        Y = spectral_model.forward(affinity_mat)
+        return Y 
 
     def forward(self, param_dict: Dict[str, torch.Tensor]) -> torch.LongTensor:
         """
@@ -1177,14 +1253,12 @@ class SpeakerClustering(torch.nn.Module):
         timestamps_in_scales = param_dict['timestamps']
         multiscale_segment_counts = param_dict['multiscale_segment_counts']
         multiscale_weights = param_dict['multiscale_weights']
-
         oracle_num_speakers = int(param_dict['oracle_num_speakers'].item())
         max_num_speakers = int(param_dict['max_num_speakers'].item())
         enhanced_count_thres = int(param_dict['enhanced_count_thres'].item())
         sparse_search_volume = int(param_dict['sparse_search_volume'].item())
         max_rp_threshold = float(param_dict['max_rp_threshold'].item())
         fixed_thres = float(param_dict['fixed_thres'].item())
-
         return self.forward_infer(
             embeddings_in_scales=embeddings_in_scales,
             timestamps_in_scales=timestamps_in_scales,
@@ -1205,8 +1279,8 @@ class SpeakerClustering(torch.nn.Module):
         multiscale_segment_counts: torch.LongTensor,
         multiscale_weights: torch.Tensor,
         oracle_num_speakers: int = -1,
-        max_rp_threshold: float = 0.15,
         max_num_speakers: int = 8,
+        max_rp_threshold: float = 0.15,
         enhanced_count_thres: int = 40,
         sparse_search_volume: int = 30,
         fixed_thres: float = -1.0,
@@ -1263,8 +1337,7 @@ class SpeakerClustering(torch.nn.Module):
                 The number of random trials for initializing k-means clustering. More trials can result in more stable clustering. The default is 1.
 
         Returns:
-            Y (LongTensor):
-                Speaker labels for the segments in the provided input embeddings.
+            (LongTensor): Speaker labels for the segments in the provided input embeddings.
         """
         self.embeddings_in_scales, self.timestamps_in_scales = split_input_data(
             embeddings_in_scales, timestamps_in_scales, multiscale_segment_counts
@@ -1284,42 +1357,19 @@ class SpeakerClustering(torch.nn.Module):
             max_num_speakers = oracle_num_speakers
 
         mat = getMultiScaleCosAffinityMatrix(
-            multiscale_weights, self.embeddings_in_scales, self.timestamps_in_scales, self.device
+            multiscale_weights=multiscale_weights, 
+            embeddings_in_scales=self.embeddings_in_scales, 
+            timestamps_in_scales=self.timestamps_in_scales, 
+            device=self.device
         )
 
-        nmesc = NMESC(
-            mat,
-            max_num_speakers=max_num_speakers,
-            max_rp_threshold=max_rp_threshold,
-            sparse_search=self.sparse_search,
-            sparse_search_volume=sparse_search_volume,
-            fixed_thres=fixed_thres,
-            nme_mat_size=self.nme_mat_size,
-            maj_vote_spk_count=self.maj_vote_spk_count,
-            parallelism=self.parallelism,
-            cuda=self.cuda,
-            device=self.device,
+        return self.forward_unit_infer(
+                mat=mat,
+                oracle_num_speakers=oracle_num_speakers,
+                max_rp_threshold=max_rp_threshold,
+                max_num_speakers=max_num_speakers,
+                sparse_search_volume=sparse_search_volume,
+                est_num_of_spk_enhanced=est_num_of_spk_enhanced,
+                kmeans_random_trials=kmeans_random_trials,
+                fixed_thres=fixed_thres,
         )
-
-        # If there are less than `min_samples_for_nmesc` segments, est_num_of_spk is 1.
-        if mat.shape[0] > self.min_samples_for_nmesc:
-            est_num_of_spk, p_hat_value = nmesc.forward()
-            affinity_mat = getAffinityGraphMat(mat, p_hat_value)
-        else:
-            nmesc.fixed_thres = max_rp_threshold
-            est_num_of_spk, p_hat_value = nmesc.forward()
-            affinity_mat = mat
-
-        # n_clusters is number of speakers estimated from spectral clustering.
-        if oracle_num_speakers > 0:
-            n_clusters = int(oracle_num_speakers)
-        elif est_num_of_spk_enhanced > 0:
-            n_clusters = int(est_num_of_spk_enhanced.item())
-        else:
-            n_clusters = int(est_num_of_spk.item())
-
-        spectral_model = SpectralClustering(
-            n_clusters=n_clusters, n_random_trials=kmeans_random_trials, cuda=self.cuda, device=self.device
-        )
-        Y = spectral_model.forward(affinity_mat)
-        return Y
