+++ nemo/collections/asr/parts/utils/online_clustering.py
@@ -33,7 +33,6 @@
 
 from typing import Dict, List, Tuple
 
-import numpy as np
 import torch
 from tqdm import tqdm
 
@@ -486,7 +485,7 @@ def get_first_arg_index(mat: torch.Tensor, label: int) -> int:
             Label which we want to find the first occuring index
 
     Returns:
-        (int) The first index of the given label
+        (int): The first index of the given label
     """
     return int(torch.where(mat == label)[0][0])
 
@@ -1196,422 +1195,3 @@ class OnlineSpeakerClustering(torch.nn.Module):
         # Match the permutation of the newly obtained speaker labels and the previous labels
         merged_clus_labels = self.match_labels(Y_merged=Y, add_new=add_new)
         return merged_clus_labels
-
-
-class LongFormSpeakerClustering(SpeakerClustering):
-    def __init__(
-        self, unit_window_len: int = 10000, sub_cluster_n: int = 50, cuda: bool = False,
-    ):
-        """
-        Initializes the speaker clustering class tailored for long-form audio, inheriting from the `SpeakerClustering` class.
-        The algorithm for long-form clustering is executed through the `forward_infer` function. Input embedding vectors are 
-        divided into smaller windows, each of size `unit_window_len`. Within every window, the clustering algorithm strives 
-        to identify `sub_cluster_n` distinct clusters. The obtained clustering labels are then expanded to match the original 
-        length of the input embeddings.
-
-        Args:
-            sub_cluster_n (int):
-                The target number of clusters to identify within each window.
-            unit_window_len (int):
-                The size of the windows in which the algorithm aims to identify `sub_cluster_n` clusters.
-            cuda (bool):
-                Boolean variable for toggling cuda availability.
-            device (torch.device):
-                Specifies the device for computation, valid values are `torch.device("cpu")`, `torch.device("cuda:<int>")`, or 
-                `torch.device("cuda")`.
-        """
-        super().__init__()
-        self.speaker_clustering = SpeakerClustering(cuda=cuda)
-        self.unit_window_len: int = unit_window_len
-        self.sub_cluster_n: int = sub_cluster_n
-        self.cuda = cuda
-
-    def unpack_labels(
-        self,
-        Y_aggr: torch.Tensor,
-        window_range_list: List[List[int]],
-        absolute_merge_mapping: List[List[torch.Tensor]],
-        org_len: int,
-    ):
-        """
-        Unpack the labels from the aggregated labels to the original labels.
-
-        Args:
-            Y_aggr (Tensor):
-                Aggregated label vector from the merged segments.
-            window_range_list (List[int]):
-                List of window ranges for each of the merged segments.
-            absolute_merge_mapping (List[Tensor]):
-                List of absolute mappings for each of the merged segments.
-                Each list element contains a list of two tensors:
-                    1. 0-th tensor is the absolute index of the bypassed segment (segments that are not changing their indexes, if any exist)
-                    2. 1-th tensor is the absolute index of the merged segment (segments that are changing their indexes, if any exist)
-            org_len (int):
-                Original length of the labels. In most cases, this is fairly large number (on the order of 10^5).
-
-        Returns:
-            Y_unpack (Tensor):
-                Unpacked labels from the aggregated labels
-        """
-        Y_unpack = torch.zeros((org_len,)).long().to(Y_aggr.device)
-        for (win_rng, abs_mapping) in zip(window_range_list, absolute_merge_mapping):
-            inferred_merged_embs = Y_aggr[win_rng[0] : win_rng[1]]
-            if len(abs_mapping[1]) > 0:
-                Y_unpack[abs_mapping[1]] = inferred_merged_embs[-1].clone()  # Merged
-                if len(abs_mapping[0]) > 0:
-                    Y_unpack[abs_mapping[0]] = inferred_merged_embs[:-1].clone()  # Bypass
-            else:
-                if len(abs_mapping[0]) > 0:
-                    Y_unpack[abs_mapping[0]] = inferred_merged_embs.clone()
-        return Y_unpack
-
-    def split_embs_to_windows(self, index: int, emb: torch.Tensor) -> Tuple[torch.Tensor, int]:
-        """
-        Splits the embedding tensor into smaller window-sized tensors based on a given index.
-        
-        Args:
-            index (int): The index of the desired window. This determines the starting point 
-                        of the window using the formula:
-                        start = self.unit_window_len * index
-            emb (torch.Tensor): The embedding tensor which needs to be split.
-
-        Returns:
-            emb_part (torch.Tensor): The window-sized tensor, which is a portion of the `emb`.
-            offset_index (int): The starting position of the window in the `emb` tensor.
-        """
-        if self.unit_window_len * (index + 1) > emb.shape[0]:
-            emb_part = emb[-1 * self.unit_window_len :]
-            offset_index = emb.shape[0] - self.unit_window_len
-        else:
-            emb_part = emb[self.unit_window_len * index : self.unit_window_len * (index + 1)]
-            offset_index = self.unit_window_len * index
-        return emb_part, offset_index
-
-    def forward_unit_infer(
-        self,
-        emb: torch.Tensor,
-        oracle_num_speakers: int = -1,
-        max_rp_threshold: float = 0.15,
-        max_num_speakers: int = 8,
-        sparse_search_volume: int = 30,
-        min_samples_for_nmesc: int = 1,
-        fixed_thres: float = -1.0,
-        cuda: bool = False,
-    ) -> torch.LongTensor:
-        """
-        This function is slimmed down version of forward_infer function in `SpeakerClustering` class.
-        Please refer to `SpeakerClustering` class for the original argument information.
-        """
-        if emb.shape[0] == 1:
-            return torch.zeros((1,), dtype=torch.int64)
-
-        mat = getCosAffinityMatrix(emb)
-        nmesc = NMESC(
-            mat,
-            max_num_speakers=max_num_speakers,
-            max_rp_threshold=max_rp_threshold,
-            sparse_search=self.sparse_search,
-            sparse_search_volume=sparse_search_volume,
-            fixed_thres=fixed_thres,
-            nme_mat_size=self.nme_mat_size,
-            maj_vote_spk_count=self.maj_vote_spk_count,
-            parallelism=self.parallelism,
-            cuda=cuda,
-            device=self.device,
-        )
-        if mat.shape[0] > min_samples_for_nmesc:
-            est_num_of_spk, p_hat_value = nmesc.forward()
-            affinity_mat = getAffinityGraphMat(mat, p_hat_value)
-        else:
-            est_num_of_spk = torch.tensor(1)
-            affinity_mat = mat
-
-        if oracle_num_speakers > 0:
-            n_clusters = int(oracle_num_speakers)
-        else:
-            n_clusters = int(est_num_of_spk.item())
-
-        spectral_model = SpectralClustering(n_clusters=n_clusters, cuda=self.cuda, device=self.device)
-        Y = spectral_model.forward(affinity_mat)
-        return Y
-
-    def forward(self, param_dict: Dict[str, torch.Tensor]) -> torch.LongTensor:
-        """
-        A function wrapper designed for performing inference using an exported script format.
-
-        Note:
-            A dictionary is used to facilitate inference with the exported jit model in the Triton server. 
-            This is done using an easy-to-understand naming convention.
-            See https://github.com/triton-inference-server/server/blob/main/docs/user_guide/model_configuration.md#special-conventions-for-pytorch-backend
-
-        Args:
-            param_dict (dict):
-                Dictionary containing the arguments for speaker clustering.
-                See `forward_infer` function for the argument information.
-
-        Returns:
-            Y (LongTensor):
-                Speaker labels for the segments in the given input embeddings.
-        """
-        embeddings_in_scales = param_dict['embeddings']
-        timestamps_in_scales = param_dict['timestamps']
-        multiscale_segment_counts = param_dict['multiscale_segment_counts']
-        multiscale_weights = param_dict['multiscale_weights']
-
-        oracle_num_speakers = int(param_dict['oracle_num_speakers'].item())
-        max_num_speakers = int(param_dict['max_num_speakers'].item())
-        enhanced_count_thres = int(param_dict['enhanced_count_thres'].item())
-        sparse_search_volume = int(param_dict['sparse_search_volume'].item())
-        max_rp_threshold = float(param_dict['max_rp_threshold'].item())
-        fixed_thres = float(param_dict['fixed_thres'].item())
-        return self.forward_infer(
-            embeddings_in_scales=embeddings_in_scales,
-            timestamps_in_scales=timestamps_in_scales,
-            multiscale_segment_counts=multiscale_segment_counts,
-            multiscale_weights=multiscale_weights,
-            oracle_num_speakers=oracle_num_speakers,
-            max_rp_threshold=max_rp_threshold,
-            max_num_speakers=max_num_speakers,
-            enhanced_count_thres=enhanced_count_thres,
-            sparse_search_volume=sparse_search_volume,
-            fixed_thres=fixed_thres,
-        )
-
-    def get_div_ceil_count(self, numer: int, denomin: int) -> int:
-        """
-        Calculates the ceiling of the division of two integers.
-        
-        Args:
-            numer (int): Numerator, the number of segments or clusters, for example.
-            denomin (int): Denominator, the number of speakers or clusters, for example.
-
-        Returns:
-            (int): The ceiling of the division of the two integers (number of chunks).
-        """
-        return int(torch.ceil(torch.tensor(numer / denomin)).item())
-
-    def long_forward_infer(
-        self,
-        embeddings_in_scales: torch.Tensor,
-        timestamps_in_scales: torch.Tensor,
-        multiscale_segment_counts: torch.LongTensor,
-        multiscale_weights: torch.Tensor,
-        oracle_num_speakers: int = -1,
-        max_rp_threshold: float = 0.15,
-        max_num_speakers: int = 8,
-        enhanced_count_thres: int = 80,
-        sparse_search_volume: int = 30,
-        fixed_thres: float = -1.0,
-        sub_cluster_n: int = 50,
-        unit_window_len: int = 10000,
-    ) -> torch.LongTensor:
-        """
-        This is forward function for long-form speaker clustering.
-        Please refer to `SpeakerClustering` class for the original argument information.
-        
-        In the `LongFormSpeakerClustering` process:
-        1. Input embeddings are divided into smaller windows of size `unit_window_len`.
-        2. Each window undergoes overclustering, resulting in `sub_cluster_n` fine-grained clusters.
-        3. These fine-grained clusters are merged to form the aggregated clustering labels `Y_aggr`.
-        4. The `unpack_labels` function is then employed to map the aggregated labels `Y_aggr` back to the 
-        original labels for all `org_len` input embeddings: `Y_unpack`.
-        
-        Args:
-            embeddings_in_scales (Tensor):
-                List containing concatenated Torch tensor embeddings across multiple scales.
-                The length of the list is equal to the number of scales.
-                Each tensor has dimensions of (Number of base segments) x (Embedding Dimension).
-            timestamps_in_scales (Tensor):
-                List containing concatenated Torch tensor timestamps across multiple scales.
-                The length of the list is equal to the number of scales.
-                Each tensor has dimensions of (Total number of segments across all scales) x 2.
-                Example:
-                    >>> timestamps_in_scales[0] = \
-                        torch.Tensor([[0.4, 1.4], [0.9, 1.9], [1.4, 2.4], ... [121.2, 122.2]])
-            multiscale_segment_counts (torch.LongTensor):
-                A Torch tensor containing the number of segments for each scale.
-                The tensor has dimensions of (Number of scales).
-                Example:
-                    >>> multiscale_segment_counts = torch.LongTensor([31, 52, 84, 105, 120])
-            multiscale_weights (torch.Tensor):
-                Multi-scale weights used when merging affinity scores.
-                Example:
-                    >>> multiscale_weights = torch.tensor([1.4, 1.3, 1.2, 1.1, 1.0])
-            oracle_num_speakers (int):
-                The number of speakers in a session as given by the reference transcript.
-            max_num_speakers (int):
-                The upper bound for the number of speakers in each session.
-            max_rp_threshold (float):
-                Limits the range of parameter search.
-                The clustering performance can vary based on this range.
-                The default value is 0.15.
-            enhanced_count_thres (int):
-                For shorter audio recordings, the clustering algorithm might not accumulate enough speaker profiles for each cluster.
-                Thus, the function `getEnhancedSpeakerCount` uses anchor embeddings (dummy representations) to mitigate the effects of cluster sparsity.
-                A value of 80 is recommended for `enhanced_count_thres`.
-            sparse_search_volume (int):
-                The number of p_values considered during NME analysis.
-                The default is 30. Lower values speed up the NME-analysis but might lead to poorer parameter estimations. Values below 20 are not recommended.
-            fixed_thres (float):
-                If a `fixed_thres` value is provided, the NME-analysis process will be skipped.
-                This value should be optimized on a development set for best results.
-                By default, it is set to -1.0, and the function performs NME-analysis to estimate the threshold.
-            kmeans_random_trials (int):
-                The number of random trials for initializing k-means clustering. More trials can result in more stable clustering. The default is 1.
-
-        Returns:
-            Y (LongTensor):
-                Speaker labels for the segments in the provided input embeddings.
-        """
-        self.unit_window_len = unit_window_len if unit_window_len is not None else self.unit_window_len
-        self.sub_cluster_n = sub_cluster_n if sub_cluster_n is not None else self.sub_cluster_n
-        if self.sub_cluster_n is None or self.unit_window_len is None:
-            raise ValueError(
-                f"sub_cluster_n ({self.sub_cluster_n}) and unit_window_len ({self.unit_window_len}) should be set."
-            )
-        elif (
-            all(v is not None for v in [self.sub_cluster_n, self.unit_window_len])
-            and self.sub_cluster_n >= self.unit_window_len
-        ):
-            raise ValueError(
-                f"sub_cluster_n ({self.sub_cluster_n}) should be smaller than unit_window_len ({self.unit_window_len})."
-            )
-
-        if self.sub_cluster_n <= max_num_speakers:
-            raise ValueError(
-                f"sub_cluster_n ({self.sub_cluster_n}) should be larger than max_num_speakers ({max_num_speakers})."
-            )
-
-        self.embeddings_in_scales, self.timestamps_in_scales = split_input_data(
-            embeddings_in_scales, timestamps_in_scales, multiscale_segment_counts
-        )
-        emb, _ = get_scale_interpolated_embs(
-            multiscale_weights, self.embeddings_in_scales, self.timestamps_in_scales, self.device
-        )
-        offset_index: int = 0
-        window_offset: int = 0
-        total_emb: List[torch.Tensor] = []
-        window_range_list: List[List[int]] = []
-        absolute_merge_mapping: List[List[torch.Tensor]] = []
-        total_window_count = self.get_div_ceil_count(numer=emb.shape[0], denomin=self.unit_window_len)
-
-        if not torch.jit.is_scripting():
-            pbar = tqdm(range(total_window_count), desc="Clustering Sub-Windows", leave=True, unit="window")
-        else:
-            pbar = range(total_window_count)
-
-        for win_index in pbar:
-            emb_part, offset_index = self.split_embs_to_windows(index=win_index, emb=emb)
-            Y_part = self.forward_unit_infer(
-                emb_part,
-                oracle_num_speakers=self.sub_cluster_n,
-                max_rp_threshold=max_rp_threshold,
-                max_num_speakers=self.sub_cluster_n,
-                sparse_search_volume=sparse_search_volume,
-                cuda=self.cuda,
-            )
-            num_to_be_merged = int(min(unit_window_len, emb_part.shape[0]) - sub_cluster_n)
-            min_count_per_cluster = self.get_div_ceil_count(
-                numer=self.sub_cluster_n, denomin=len(torch.unique(Y_part))
-            )
-
-            # We want only one embedding vector for each cluster, so we calculate the number of embedding vectors to be removed
-            class_target_vol = get_merge_quantity(
-                num_to_be_removed=num_to_be_merged,
-                pre_clus_labels=Y_part,
-                min_count_per_cluster=min_count_per_cluster,
-            )
-            if not torch.jit.is_scripting():
-                pbar.update(1)
-
-            # `class_target_vol` is a list of cluster-indices from overclustering
-            for spk_idx, merge_quantity in enumerate(list(class_target_vol)):
-                merged_embs, merged_clus_labels, index_mapping = run_reducer(
-                    pre_embs=emb_part, target_spk_idx=spk_idx, merge_quantity=merge_quantity, pre_clus_labels=Y_part,
-                )
-                total_emb.append(merged_embs)
-                absolute_index_mapping = [x + offset_index for x in index_mapping]
-                absolute_merge_mapping.append(absolute_index_mapping)
-                window_range_list.append([window_offset, window_offset + merged_embs.shape[0]])
-                window_offset += merged_embs.shape[0]
-
-        if not torch.jit.is_scripting():
-            pbar.close()
-
-        # Concatenate the reduced embeddings then perform high-level clustering
-        reduced_embs = torch.cat(total_emb)
-        Y_aggr = self.forward_unit_infer(
-            reduced_embs,
-            oracle_num_speakers=oracle_num_speakers,
-            max_rp_threshold=max_rp_threshold,
-            max_num_speakers=max_num_speakers,
-            sparse_search_volume=sparse_search_volume,
-            fixed_thres=fixed_thres,
-            cuda=self.cuda,
-        )
-
-        Y_unpack = self.unpack_labels(
-            Y_aggr=Y_aggr,
-            window_range_list=window_range_list,
-            absolute_merge_mapping=absolute_merge_mapping,
-            org_len=emb.shape[0],
-        )
-        if reduced_embs.shape[0] != Y_aggr.shape[0]:
-            raise ValueError(
-                "The number of embeddings (reduced_embs.shape) and the number of labels (Y_aggr.shape) do not match."
-            )
-        return Y_unpack
-
-    def forward_infer(
-        self,
-        embeddings_in_scales: torch.Tensor,
-        timestamps_in_scales: torch.Tensor,
-        multiscale_segment_counts: torch.LongTensor,
-        multiscale_weights: torch.Tensor,
-        oracle_num_speakers: int = -1,
-        max_rp_threshold: float = 0.15,
-        max_num_speakers: int = 8,
-        enhanced_count_thres: int = 80,
-        sparse_search_volume: int = 30,
-        fixed_thres: float = -1.0,
-        sub_cluster_n: int = 50,
-        unit_window_len: int = 10000,
-    ) -> torch.LongTensor:
-        """
-        This function is a wrapper designed for toggling between long-form and short-form speaker clustering.
-        The details of short-form clustering is in `SpeakerClustering` class.
-        Note that `torch.jit.script` currently does not support `**kwargs` in the function signature therefore,
-        we need to use a wrapper function to handle the arguments.
-        """
-        self.unit_window_len = unit_window_len if unit_window_len is not None else self.unit_window_len
-        self.sub_cluster_n = sub_cluster_n if sub_cluster_n is not None else self.sub_cluster_n
-        if embeddings_in_scales.shape[0] > self.unit_window_len:
-            return self.long_forward_infer(
-                embeddings_in_scales=embeddings_in_scales,
-                timestamps_in_scales=timestamps_in_scales,
-                multiscale_segment_counts=multiscale_segment_counts,
-                multiscale_weights=multiscale_weights,
-                oracle_num_speakers=oracle_num_speakers,
-                max_rp_threshold=max_rp_threshold,
-                max_num_speakers=max_num_speakers,
-                enhanced_count_thres=enhanced_count_thres,
-                sparse_search_volume=sparse_search_volume,
-                fixed_thres=fixed_thres,
-                sub_cluster_n=sub_cluster_n,
-                unit_window_len=unit_window_len,
-            )
-        else:
-            cluster_labels = self.speaker_clustering.forward_infer(
-                embeddings_in_scales=embeddings_in_scales,
-                timestamps_in_scales=timestamps_in_scales,
-                multiscale_segment_counts=multiscale_segment_counts,
-                multiscale_weights=multiscale_weights,
-                oracle_num_speakers=oracle_num_speakers,
-                max_rp_threshold=max_rp_threshold,
-                max_num_speakers=max_num_speakers,
-                enhanced_count_thres=enhanced_count_thres,
-                sparse_search_volume=sparse_search_volume,
-                fixed_thres=fixed_thres,
-            )
-            self.timestamps_in_scales = self.speaker_clustering.timestamps_in_scales
-            return cluster_labels
